separate searching from getting box office (see #3)
This splits the current API in 2: /search to look up movies by title,
and /boxoffice to look up box office results once you have a movie ID.
The previous endpoint is preserved (for now) as /boxoffice/search since
that is still what the UI uses.

The next step will be to update the UI to support searching by movie
title first.